requirements.txt
Brotli==1.1.0
coverage==7.4.0
elasticsearch==7.16.2
networkx==3.1
numpy==1.24.4
pre-commit==3.5.0
pygraphviz==1.11
pytest==7.4.4
python-dotenv==0.15.0
pytz==2019.1
pyyaml>=6.0
pyzipper==0.3.6
r2pipe==1.8.8
requests==2.21.0
rich>=13.0.0
urllib3==1.24.3
